Clarified part of the Statements and Control Flow
The description said that a semicolon separator was "entirely optional." The code below used colons. I checked with a more knowledgeable person and they said that the semicolon shouldn't be used in the place of the colon and that the colon was required. I made changes to the wording to reflect this in the GDScript Basics documentation.